### 2021-07-22 #3327 @leouieda
👋🏽 hi @joergbuchwald this submission is a little bit unusual since it is reporting on two separate software packages. I've been reading through your paper and project documentation to get a better idea for both of them. Could you please briefly explain your reasoning for including both of them in a single paper? To be clear, this is unusual but not prohibited so I'm just trying to better understand your paper.
