### 2021-07-22 #3327 @joergbuchwald
Hi @leouieda thank you for the question. Both packages evolved together as pre- and post-processing utils for OGS. However, the application field of VTUinterface is somewhat broader, as also other simulation software are using VTU and PVD as output formats, so it made sense for us to have separate packages. For this paper, the original idea was to sell the pre- and postprocessing of OGS in python as one workflow to demonstrate what can be done with OGS in one single jupyter notebook. I also think, the advantages of using ogs6py for people who are used to using paraview for data analysis might not be obvious at first sight. It really comes with simple access to VTU and PVD data.
Which is why we wanted to present an entire workflow as demonstrated in this (jupyter notbook)[https://github.com/joergbuchwald/ogs6py/blob/master/paper/paper_ogs6py_vtuio.ipynb]. However, this was too long for the JOSS word limit, so we needed to shorten the paper and therefore the original idea got possibly also a bit lost.
