### 2021-08-26 1d4e118
put JOSS paper into separate repo
