+++ paper/paper.md
@@ -1,0 +1,169 @@
+---
+title: 'ogs6py and VTUinterface: streamlining OpenGeoSys workflows in Python'
+tags:
+  - Python
+  - physics
+  - THMC
+  - VTU
+  - time-series
+  - sensitivity analysis
+  - uncertainty quantification
+  - OpenGeoSys
+
+authors:
+  - name: Jörg Buchwald^[corresponding author]
+    orcid: 0000-0001-5174-3603
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Olaf Kolditz
+    orcid: 0000-0002-8098-4905
+    affiliation: "1, 3, 4"
+  - name: Thomas Nagel
+    orcid: 0000-0001-8459-4616
+    affiliation: "2, 4"
+affiliations:
+ - name: Helmholtz Center for Environmental Research - UFZ, Leipzig, Germany
+   index: 1
+ - name: Technische Universität Bergakademie Freiberg, Germany
+   index: 2
+ - name: Technische Universität Dresden, Germany
+   index: 3
+ - name: TUBAF-UFZ Center for Environmental Geosciences, Germany
+   index: 4
+date: 12 May 2021
+bibliography: paper.bib
+
+---
+
+# Summary
+
+[ogs6py](https://github.com/joergbuchwald/ogs6py) is a Python interface for the open-source package OpenGeoSys [@Bilke2019], a finite element code
+for the simulation of multi-field processes in fractured porous media.
+In conjunction with [VTUinterface](https://github.com/joergbuchwald/VTUinterface) it is possible to streamline modeling workflows
+in Jupyter Notebooks using Python.
+With this article, we wish to introduce two new Python modules that facilitate
+the pre- and post-processing of finite element calculations of OpenGeoSys and thus
+make this code more accessible to the community. Their use is demonstrated along workflows typically
+encountered by the modeler, including the variation of parameters, boundary conditions,
+or solver settings, the verification of simulation results by comparison to analytical
+solutions, the set-up and evaluation of ensemble runs, the rapid analysis of results by line plots,
+time series, or transient contour plots.
+
+# Statement of need
+
+Driven by its ease-of-use and flexibility as an open-source 
+dynamic language, its vast modular ecosystem, the development of powerful plotting
+libraries and the Jupyter Notebook technology, Python became a widely used common framework for 
+scientific data analysis in the modeling community during the past decade.
+However, the attractiveness of Python is not just limited to post-processing. 
+E.g., with the Python wrapper for GMSH [@geuzaine2009gmsh] or the tool meshio [@nico_schlomer_2021_4745399] also pre-processing tasks can
+be easily conducted without leaving the IPython command prompt. The usability of any modeling package
+is therefore greatly enhanced if Python bindings are provided. In fact, 
+while many open-source tools effectively forced the user to learn a singular syntax
+for interacting with the software, Python bindings allow control over such tools from 
+within the Python world and thus open them up for a wider community of users.
+
+Here, we are particularly addressing the open-source code OpenGeoSys (OGS) [@Bilke2019] version 6. It is our aim
+to facilitate both pre- and post-processing workflows utilizing the Python ecosystem. 
+This aim was not the least inspired by the desire to facilitate setting up, controlling and
+evaluating ensemble runs [@Buchwald2020;@Chaudhry2021] but has now taken on a wider perspective of general 
+software usability. There exists already a similar Python interface "ogs5py" for OGS version 5 [@muller2021ogs5py]. 
+However, the differences in many concepts, like the use of XML input files, required an entirely new package to be built from scratch.
+
[…]
