+++ demo/paper_ogs6py_vtuio.md
@@ -1,0 +1,547 @@
+# Summary
+
+ogs6py is a Python interface for the OpenGeoSys finite element software [@Bilke2019].
+In conjunction with VTUinterface it is possible to streamline modeling workflows
+in Jupyter Notebooks using Python.
+With this article, we wish to introduce two new Python modules that facilitate
+the pre- and postprocessing of finite element calculations of OpenGeoSys and thus
+make this code more accessible. Their use is demonstrated along workflows typically
+encountered by the modeller, including the variation of parameters, boundary conditions,
+or solver settings, the verification of simulation results by comparison to analytical
+solutions, the set-up and evaluation of ensemble runs, the analysis of results by line plots,
+time series, or transient contour plots.
+
+# Statement of need
+
+Driven by its ease-of-use and flexibility as an open-source 
+dynamic language, its vast modular ecosystem, the development of powerful plotting
+libraries and the Jupyter Notebook technology, Python became the quasi-standard for 
+scientific data analysis in the modelling community during the past decade.
+However, the attractiveness of Phython is not just limited to postprocessing. 
+E.g, with the Python wrapper for GMSH [@geuzaine2009gmsh] or the tool meshio [@nico_schlomer_2021_4745399] also pre-processing tasks can
+be easily conducted without leaving the IPython command prompt. The useability of a modeling package
+is therefore greatly enhanced if Python bindings are provided. In fact, 
+while many open-source tools effectively forced the user to learn a singular syntax
+for interacting with the software, Python bindings allow control over such tools from 
+within the Python world and thus open them up for a wider community of users.
+
+Here, we are particularly addressing the open-source code OpenGeoSys (OGS) [@Bilke2019] version 6. It is our aim
+to facilitate both pre- and post-processing workflows using the Python ecosystem. 
+This aim was not the least inspired by the desire to facilitate setting up, controlling and
+evaluating ensemble runs [@Buchwald2020,@Chaudhry2021] but has now taken on a wider perspective of general 
+software usability. There exists already a similar python interface "ogs5py" for OGS version 5 [@muller2021ogs5py]. 
+However, the differences in many concepts, like input file handling, required an entirely new package build from scratch.
+
+As standard output format, OpenGeoSys uses VTK unstructured grid files (VTU) as timeslices stacked together by a PVD file.
+These can be analyzed typically using Paraview [@ahrens2005paraview]. For interactive Python use the Python 
+wrapper for VTK [@schroeder2000visualizing] and some other tools like PyVista [@sullivan2019pyvista] or Mayavi [@ramachandran2011mayavi] 
+are available facilitating an easier access to the VTK library.
+While the direct use of the VTK library is quite cumbersome for quite _simple_ tasks, like reading data for a given point set, especially when interpolation between grid points is also required. The latter packages focus mainly on 3D visualization. However, the _bread and butter_ bussiness of a finite-element-modeller often cosists of the extraction of single- or multiple point time-series data.
+To our knowledge the named packages (with the exception of Paraview) don't have file support for PVDs or time series data, yet ([@pvdissue; @timeseriesissue].
+
+# Usage
+
[…]
