+++ README.md
@@ -1,1 +1,7 @@
-# joss_ogs6py_VTUinterface
+# JOSS paper ogs6py and VTUinterface: streamlining OpenGeoSys workflows in Python
+
+[VTUinterface](https://github.com/joergbuchwald/VTUinterface)
+
+[ogs6py](https://github.com/joergbuchwald/ogs6py)
+
+

[…]
