+++ README.md
@@ -4,4 +4,6 @@
 
 [ogs6py](https://github.com/joergbuchwald/ogs6py)
 
+[paper](https://github.com/joergbuchwald/joss_ogs6py_VTUinterface/blob/master/paper/paper.md)
 
+[Workflow demonstration (jupyter notebook)](https://github.com/joergbuchwald/joss_ogs6py_VTUinterface/blob/master/demo/paper_ogs6py_vtuio.ipynb)
