### 2021-08-27 cb98370
Update paper.md
combine sentence / fix grammar
