### 2021-08-31 #3673 @whedon
[REVIEW]: ogs6py and VTUinterface: streamlining OpenGeoSys workflows in Python
**Submitting author:** @joergbuchwald (<a href="http://orcid.org/0000-0001-5174-3603">Jörg Buchwald</a>)
**Repository:** <a href="https://github.com/joergbuchwald/joss_ogs6py_VTUinterface" target ="_blank">https://github.com/joergbuchwald/joss_ogs6py_VTUinterface</a>
**Version:** ogs6py-v0.34/VTUinterface-v0.69

@cpgr & @TobbeTripitaka & @akaszynski , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
