### 2021-08-31 #3673 @leouieda
:wave: @cpgr, @TobbeTripitaka, @akaszynski @joergbuchwald  this is the review thread for the paper. All of our communications will happen here from now on.

All reviewers have checklists at the top of this thread with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

We aim for reviews to be completed within about 4 weeks. Please let me know if any of you require some more time, we are happy to make acomodations. We can also use @whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Please feel free to ping me here (@leouieda) or email me privately if you have any questions/concerns.
