### 2021-09-06 53f4c31
Some suggestions to improve readability
I made some suggestions to improve readability, and some idiomatic corrections. 

Some edits are rather far-reaching and extensive. Please, confirm that my suggestions do not change the intended meaning.
