+++ paper/paper.md
@@ -36,62 +36,32 @@ bibliography: paper.bib
 
 # Summary
 
-[ogs6py](https://github.com/joergbuchwald/ogs6py) is a Python interface for the open-source package OpenGeoSys [@Bilke2019], a finite element code
-for the simulation of multi-field processes in fractured porous media.
-In conjunction with [VTUinterface](https://github.com/joergbuchwald/VTUinterface) it is possible to streamline modeling workflows
-in Jupyter Notebooks using Python.
-With this article, we wish to introduce two new Python modules that facilitate
-the pre- and post-processing of finite element calculations of OpenGeoSys and thus
-make this code more accessible to the community. Their use is demonstrated along workflows typically
-encountered by the modeler, including the variation of parameters, boundary conditions,
-or solver settings, the verification of simulation results by comparison to analytical
-solutions, the set-up and evaluation of ensemble runs, the rapid analysis of results by line plots,
-time series, or transient contour plots.
+We introduce two new Python modules that facilitate the pre- and post-processing of finite element calculations. [ogs6py](https://github.com/joergbuchwald/ogs6py) is a Python interface for the open-source package OpenGeoSys [@Bilke2019], a finite element code for simulation of multi-field processes in fractured porous media. Modeling workflows is further streamlined in Jupyter Notebooks using the newly developed [VTUinterface](https://github.com/joergbuchwald/VTUinterface).
+The usage of the modules is demonstrated with common workflow operations, including parameter variations, boundary conditions, solver settings, verification of simulation results by comparison to analytical solutions, set-up and evaluation of ensemble runs, convenient analysis of results by line plots, time series, or transient contour plots.
 
 # Statement of need
 
-Driven by its ease-of-use and flexibility as an open-source 
-dynamic language, its vast modular ecosystem, the development of powerful plotting
-libraries and the Jupyter Notebook technology, Python became a widely used common framework for 
-scientific data analysis in the modeling community during the past decade.
-However, the attractiveness of Python is not just limited to post-processing. 
-E.g., with the Python wrapper for GMSH [@geuzaine2009gmsh] or the tool meshio [@nico_schlomer_2021_4745399] also pre-processing tasks can
-be easily conducted without leaving the IPython command prompt. The usability of any modeling package
-is therefore greatly enhanced if Python bindings are provided. In fact, 
-while many open-source tools effectively forced the user to learn a singular syntax
-for interacting with the software, Python bindings allow control over such tools from 
-within the Python world and thus open them up for a wider community of users.
+Python has become a widely used framework for scientific data analysis and modeling. The development is driven by ease of use and flexibility, the vast modular ecosystem including powerful plotting libraries, and the Jupyter Notebook technology. The attractiveness of Python is not limited to post-processing; pre-processing tasks can simply be conducted, using packages as the Python wrapper for GMSH [@geuzaine2009gmsh] and the tool meshio [@nico_schlomer_2021_4745399]. 
+While many existing open-source tools force the user to learn a new syntax for interacting with the software, Python bindings allow control in a general language and thus more accessible for a wider community of users.
 
-Here, we are particularly addressing the open-source code OpenGeoSys (OGS) [@Bilke2019] version 6. It is our aim
-to facilitate both pre- and post-processing workflows utilizing the Python ecosystem. 
-This aim was not the least inspired by the desire to facilitate setting up, controlling and
-evaluating ensemble runs [@Buchwald2020;@Chaudhry2021] but has now taken on a wider perspective of general 
[…]
