### 2021-09-06 46733d6
Update paper.md
small fixes
