+++ paper/paper.md
@@ -117,6 +117,9 @@ plt.ylabel("$p$ / MPa");
 
 This brief overview shows only some of the functionalities coming with ogs6py and VTUinterface. Further developments will focus on extending functionalities focusing on built-in checks to ensure that only valid input files are generated.
 
+# Technical Details
+
+ogs6py requires python 3.9 and uses [lxml](https://lxml.de/) to process OGS6 input files and uses the subprocess module to run OGS. Furthermore [pandas](https://pandas.pydata.org/) is required to for holding OGS logfile data. VTUinterface requires python 3.8 and uses the [vtk](https://vtk.org/) wrapper for python to access VTU files and lxml for PVD files. In addition to vtk's own interpolation functionalities, we use pandas and scipy for interpolation.
 
 # Applications
 
