### 2021-09-15 #3673 @cpgr
I notice that the paper has been updated since I first read it
