### 2021-09-15 #3673 @akaszynski
Raised an issue regarding `ogs6py`: https://github.com/joergbuchwald/ogs6py/issues/14
Paper reads fairly well.  Made minor edits here: https://github.com/joergbuchwald/joss_ogs6py_VTUinterface/pull/3
