### 2021-09-15 #3673 @akaszynski
> @akaszynski thanks for the review comments! I see a few unchecked item in your checklist and I assume they’re related to the issues you raised. Please go ahead and mark them as done things are resolved.

Will do.  Busy with day job, will work on it this evening.
