### 2021-09-22 joergbuchwald/joss_ogs6py_VTUinterface#4 @cpgr
Incorrect filename in paper_ogs6py_vtuio.ipynb
The line where the OGS project file is opened in the demo notebook
```
model = OGS(INPUT_FILE="tunnel_ogs6py.prj", PROJECT_FILE="tunnel_exc.prj", MKL=True)
```
refers to a `INPUT_FILE` that doesn't exist in the repository. I think it should just be `INPUT_FILE="tunnel.prj"`.

Interestingly, that cell executes without error, even though the input file doesn't exist. It is only when executing the following cell in the notebook
```
model.replace_text("tunnel_exc", xpath="./time_loop/output/prefix")
```
that an error is thrown
```
(lots of stack traces here)
OSError: Error reading file 'tunnel_ogs6py.prj': failed to load external entity "tunnel_ogs6py.prj"
```

Ref openjournals/joss-reviews#3673
