### 2021-09-22 joergbuchwald/joss_ogs6py_VTUinterface#4 @cpgr
Note: this example is also listed in the JOSS paper, so make sure that this works out of the box!
