### 2021-09-22 joergbuchwald/joss_ogs6py_VTUinterface#4 @joergbuchwald
Yes, it is `tunnel.prj`. I corrected this and also a method call somewhere below that has changed meanwhile.
Probably, it is a good idea to check whether the filename exists already when the object is instantiated.
