### 2021-09-22 joergbuchwald/joss_ogs6py_VTUinterface#4 @cpgr
I think checking if the file exists during instantiation is probably a good idea.

Don't forget to change paper.md to `tunnel.prj` as well in the example.
