### 2021-09-23 joergbuchwald/joss_ogs6py_VTUinterface#5 @cpgr
Suggestions for minor edits to paper.md
I also have some suggestions for `paper.md` to improved readability (my final issue I promise :smile:)

```diff
Summary

-The usage of the modules is demonstrated with common workflow operations, including parameter variations, boundary
-conditions, solver settings, verification of simulation results by comparison to analytical solutions, set-up and evaluation of
-ensemble runs, convenient analysis of results by line plots, time series, or transient contour plots.
+The use of the modules is demonstrated with common workflow operations, including parameter variations, setting
+boundary conditions, changing solver settings, verification of simulation results by comparison to analytical solutions,
+set-up and evaluation of ensemble runs, and convenient analysis of results by line plots, time series, or transient contour
+plots.

Statement of need
-The attractiveness of Python is not limited to post-processing; pre-processing tasks can simply be conducted, using
-packages as the Python wrapper for GMSH [@geuzaine2009gmsh] and the tool meshio [@nico_schlomer_2021_4745399].
-While many existing open-source tools force the user to learn a new syntax for interacting with the software, Python
-bindings allow control in a general language and thus more accessible for a wider community of users.
+The attractiveness of Python is not limited to post-processing; pre-processing tasks can be simply conducted,
+using packages such as the Python wrapper for GMSH [@geuzaine2009gmsh] or the tool meshio
+[@nico_schlomer_2021_4745399]. While many existing open-source tools force the user to learn a new syntax for
+interacting with the software, Python bindings allow control in a general language and thus are more accessible for a wider
+community of users.

-The mentioned Python packages focus on visualization aspects, and except for Paraview, to our knowledge, the mentione
-packages do not have file support for PVD files or time-series data[@pvdissue; @timeseriesissue].
+The mentioned Python packages focus on visualization aspects, and except for Paraview, to our knowledge, do not have file
+support for PVD files or time-series data[@pvdissue; @timeseriesissue].

Features
-model = OGS(INPUT_FILE="tunnel_ogs6py.prj", PROJECT_FILE="tunnel_exc.prj")
+model = OGS(INPUT_FILE="tunnel.prj", PROJECT_FILE="tunnel_exc.prj")

-One of the most significant features of VTUinterface is the ability to deal with PVD files as time-series data.
-E.g., the following command reads in the VTU point field “pressure” at point “pt0,” defined in a dictionary, using nearest
-neighbour interpolation.
+One of the most powerful features of VTUinterface is the ability to deal with PVD files as time-series data.
+For example, the following command reads in the VTU point field “pressure” at point “pt0,” defined in a dictionary, using nearest
+neighbour interpolation.

Technical details
-ogs6py requires python 3.9 and uses lxml to process OGS6 input files and uses the subprocess module to run OGS.
+ogs6py requires python 3.8 or above and uses lxml to process OGS6 input files and uses the subprocess module to run OGS.

-VTUinterface requires python 3.8 ...
+VTUinterface requires python 3.8 or above ...

Applications
-Both introduced packages are relatively new, being only 1 to 2 years old.
+Both of the packages introduced are relatively new, being only 1 to 2 years old.
-E.g., a YouTube video was published explaining their use;
+For example, a YouTube video was published explaining their use;
```

**Note** I think I tested `ogs6py` in python 3.8, so I think it works with that (at least, that is what you changed the min version to in #4. So I suggested that change in the paper above.

Congratulations again on developing these tools. I think that lots of OGS users will find them very helpful.

Refs openjournals/joss-reviews#3673
