+++ paper/paper.md
@@ -37,28 +37,26 @@ bibliography: paper.bib
 # Summary
 
 We introduce two new Python modules that facilitate the pre- and post-processing of finite element calculations. [ogs6py](https://github.com/joergbuchwald/ogs6py) is a Python interface for the open-source package OpenGeoSys [@Bilke2019], a finite element code for simulation of multi-field processes in fractured porous media. Modeling workflows can be further streamlined in Jupyter Notebooks using the newly developed [VTUinterface](https://github.com/joergbuchwald/VTUinterface).
-The usage of the modules is demonstrated with common workflow operations, including parameter variations, boundary conditions, solver settings, verification of simulation results by comparison to analytical solutions, set-up and evaluation of ensemble runs, convenient analysis of results by line plots, time series, or transient contour plots.
+The use of the modules is demonstrated with common workflow operations, including parameter variations, setting boundary conditions, changing solver settings, verification of simulation results by comparison to analytical solutions, set-up and evaluation of ensemble runs, and convenient analysis of results by line plots, time series, or transient contour plots.
 
 # Statement of need
 
-Python has become a widely used framework for scientific data analysis and modeling. The development is driven by ease of use and flexibility, the vast modular ecosystem including powerful plotting libraries, and the Jupyter Notebook technology. The attractiveness of Python is not limited to post-processing; pre-processing tasks can simply be conducted, using packages as the Python wrapper for GMSH [@geuzaine2009gmsh] and the tool meshio [@nico_schlomer_2021_4745399]. 
-While many existing open-source tools force the user to learn a new syntax for interacting with the software, Python bindings allow control in a general language and thus more accessible for a wider community of users.
+Python has become a widely used framework for scientific data analysis and modeling. The development is driven by ease of use and flexibility, the vast modular ecosystem including powerful plotting libraries, and the Jupyter Notebook technology. The attractiveness of Python is not limited to post-processing; pre-processing tasks can be simply conducted, using packages such as the Python wrapper for GMSH [@geuzaine2009gmsh] or the tool meshio [@nico_schlomer_2021_4745399]. While many existing open-source tools force the user to learn a new syntax for interacting with the software, Python bindings allow control in a general language and thus are more accessible for a wider community of users.
 
 In this contribution, we address interaction with the open-source code OpenGeoSys (OGS) [@Bilke2019] version 6, aiming to facilitate both pre-and post-processing workflows with Python. This aim was partly inspired by the desire to design, control and evaluate ensemble runs [@Buchwald2020;@Chaudhry2021] but has now taken on a wider perspective for general usability. A similar Python interface "ogs5py" exists for OGS version 5 [@muller2021ogs5py]; however, conceptual differences between the versions, for example, the use of XML input files, required an entirely new package to be built from scratch.
 
[…]
