### 2021-09-23 joergbuchwald/joss_ogs6py_VTUinterface#4 @joergbuchwald
> I think checking if the file exists during instantiation is probably a good idea.
I added now file existence checks to both packages.
