### 2021-10-05 #3673 @TobbeTripitaka
Some edits were suggested, and mostly incoperated in the manuscript. https://github.com/joergbuchwald/joss_ogs6py_VTUinterface/pull/1
