+++ paper/paper.bib
@@ -174,16 +174,16 @@ year = {2019}
 }
 
 @misc{pvdissue,
-  author = {banesullivan},
-  title = {pyvista issue: Add .pvd reader},
+  author = {Sullivan, C},
+  title = {PyVista issue 414: Add .pvd reader},
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository issue},
   url = {https://github.com/pyvista/pyvista/issues/414}
 }
 @misc{timeseriesissue,
-  author = {awa5114},
-  title = {pyvista issue: Time series data support in pyvista},
+  author = {Aboufirass, A},
+  title = {PyVista issue 294: Time series data support in pyvista},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository issue},
