### 2021-11-16 #3673 @leouieda
@joergbuchwald @nagelt I'm very happy with the state of the paper and the reviewers are satisfied as well so I'm happy to move on to acceptance of your paper 🎉

There just a few of things that you'll need to do now:

1. Please merge the minor fixes in joergbuchwald/joss_ogs6py_VTUinterface#7
1. Double check authors and affiliations (including ORCIDs)
1. Make a release of both software packages with the latest changes from the review. This is the version that will be used in the JOSS paper.  **Post the version numbers here** once they are available.
1. Publish the 2 releases **and** the paper repository on a **single** Zenodo/figshare/etc archive and post the DOI here.
1. Make sure that the **title and author list** (including ORCIDs) in the Zenodo/figshare archive match those in the JOSS paper.

Once those are done, we can move on to publication! Let me know if you need any help with these or have any questions.
