### 2021-11-16 #3673 @joergbuchwald
Thank you very much for editing :)

ogs6py is now in version 0.34 and VTUinterface in version 0.69.
I published everything together now under zenodo DOI 10.5281/zenodo.5705727.
