### 2021-11-19 #3673 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
