### 2021-11-19 #3673 @arfon
@joergbuchwald  – there are a number of software packages mentioned in the paper (e.g., Jupyter, Pandas, scipy) which aren't cited. Could you please look up the default citations for these packages and cite them in your paper?

Many thanks!
