### 2021-11-20 a718ece
add missing citations
