+++ paper/paper.bib
@@ -1,3 +1,41 @@
+@book{kluyver2016jupyter,
+  title={Jupyter Notebooks-a publishing format for reproducible computational workflows.},
+  author={Kluyver, Thomas and Ragan-Kelley, Benjamin and P{\'e}rez, Fernando and Granger, Brian E and Bussonnier, Matthias and Frederic, Jonathan and Kelley, Kyle and Hamrick, Jessica B and Grout, Jason and Corlay, Sylvain and others},
+  volume={2016},
+  year={2016}
+}
+@article{virtanen2020scipy,
+  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
+  journal={Nature methods},
+  volume={17},
+  number={3},
+  pages={261--272},
+  year={2020},
+  publisher={Nature Publishing Group}
+}
+@inproceedings{mckinney2010data,
+  title={Data structures for statistical computing in python},
+  author={McKinney, Wes and others},
+  booktitle={Proceedings of the 9th Python in Science Conference},
+  volume={445},
+  pages={51--56},
+  year={2010},
+  organization={Austin, TX}
+}
+@misc{behnel2005lxml,
+  title={lxml: XML and HTML with Python},
+  author={Behnel, Stefan and Faassen, Martijn and Bicking, Ian},
+  year={2005},
+  publisher={Lxml}
+}
+@book{VTK,
+  author={Will Schroeder and Ken Martin and Bill Lorensen},
+  title={{The Visualization Toolkit--An Object-Oriented Approach To 3D Graphics}},
+  publisher={Kitware, Inc.},
+  edition={Fourth},
+  year={2006}
+}
 @article{Wang2021,
 author = {Wang, Wenqing and Shao, Hua and Nagel, Thomas and Kolditz, Olaf},
 doi = {10.1016/j.ijrmms.2021.104683},
