+++ paper/paper.bib
@@ -2,6 +2,7 @@
   title={Jupyter Notebooks-a publishing format for reproducible computational workflows.},
   author={Kluyver, Thomas and Ragan-Kelley, Benjamin and P{\'e}rez, Fernando and Granger, Brian E and Bussonnier, Matthias and Frederic, Jonathan and Kelley, Kyle and Hamrick, Jessica B and Grout, Jason and Corlay, Sylvain and others},
   volume={2016},
+  doi={10.3233/978-1-61499-649-1-87},
   year={2016}
 }
 @article{virtanen2020scipy,
@@ -12,6 +13,7 @@
   number={3},
   pages={261--272},
   year={2020},
+  doi={10.1038/s41592-019-0686-2},
   publisher={Nature Publishing Group}
 }
 @inproceedings{mckinney2010data,
@@ -21,6 +23,7 @@
   volume={445},
   pages={51--56},
   year={2010},
+  doi={10.25080/Majora-92bf1922-00a},
   organization={Austin, TX}
 }
 @misc{behnel2005lxml,
