### 2021-11-20 #3673 @joergbuchwald
@arfon thanks for the hint. I added now all missing citations.
